requirements.txt
# LM Studio Translator Dependencies

# Requires Python 3.9+ (argparse.BooleanOptionalAction)

# Core dependencies
openai>=1.55.0

# GUI dependencies
# tkinter - Required for GUI interface (Python standard library, usually pre-installed)
# Note: On some Linux distributions, you may need to install tkinter separately:
#   Ubuntu/Debian: sudo apt-get install python3-tk
#   CentOS/RHEL/Fedora: sudo yum install tkinter or sudo dnf install python3-tkinter

# Modern GUI theming
# ttkbootstrap removed

# Document format parsing
pdfminer.six>=20221105
PyPDF2>=3.0.1
python-docx>=1.1.0
PyMuPDF>=1.24.9
requests>=2.32.0
ttkbootstrap>=1.10.1

# Optional development dependencies
# pytest>=8.0.0